agentic_ai/tousif/traveL_bot/pyproject.toml
[project]
name = "travel-advisory-bot"
version = "0.1.0"
description = "Agentic AI Travel Advisory Bot with multimodal input support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.29.0",
    "openai>=1.0.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0",
    "PyPDF2>=3.0.0",
    "python-multipart>=0.0.6",
    "langdetect>=1.0.9",
    "googletrans>=4.0.0",
    "opencv-python>=4.8.0",
    "whisper-openai>=20231117",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "folium>=0.15.0",
    "plotly>=5.17.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0"
]
